+++ cda.sh
@@ -30,7 +30,7 @@ _cda()
     #------------------------------------------------
     # app info
     declare -r APPNAME="cda"
-    declare -r VERSION="1.3.0 (2018-03-19)"
+    declare -r VERSION="1.4.0 (2018-05-28)"
 
     # override system variables
     local IFS=$' \t\n'
@@ -53,9 +53,7 @@ _cda()
     declare -r LIST_DIR="$CDA_DATA_ROOT/$LIST_DIR_NAME"
     declare -r LIST_DEFAULT_NAME="default"
     declare -r LIST_NAME_FILE="$CDA_DATA_ROOT/listname"
-    declare -r AUTO_ALIAS_LIST_NAME=".autolist"
-    declare -r AUTO_ALIAS_LIST_FILE="$LIST_DIR/$AUTO_ALIAS_LIST_NAME"
-    declare -r AUTO_ALIAS_PWD_FILE="$LIST_DIR/.autopwd"
+    declare -r LAST_PATH_FILE="$LIST_DIR/.lastpath"
 
     # flags
     declare -r FLAG_NONE=0
@@ -80,13 +78,11 @@ _cda()
     declare -r FLAG_LIST_PATHS=$((1<<18))
     declare -r FLAG_CHECK=$((1<<19))
     declare -r FLAG_CLEAN=$((1<<20))
-    declare -r FLAG_MULTIPLEXER=$((1<<21))
-    declare -r FLAG_OVERWRITE=$((1<<22))
-    declare -r FLAG_FILTER_FORCED=$((1<<23))
-    declare -r FLAG_SUBDIR=$((1<<24))
-    declare -r FLAG_UPDATE_AUTOALIAS=$((1<<25))
-    declare -r FLAG_BUILTIN_CD=$((1<<26))
-    declare -r FLAG_VERBOSE=$((1<<27))
+    declare -r FLAG_OVERWRITE=$((1<<21))
+    declare -r FLAG_FILTER_FORCED=$((1<<22))
+    declare -r FLAG_SUBDIR=$((1<<23))
+    declare -r FLAG_BUILTIN_CD=$((1<<24))
+    declare -r FLAG_VERBOSE=$((1<<25))
     
     declare -r FLAGS_INCOMPAT=$((
           FLAG_HELP_SHORT | FLAG_HELP_FULL | FLAG_VERSION
@@ -94,20 +90,18 @@ _cda()
         | FLAG_EDIT_LIST | FLAG_EDIT_CONFIG | FLAG_PATH
         | FLAG_LIST_NAMES | FLAG_LIST_PATHS | FLAG_LIST
         | FLAG_LIST_FILES | FLAG_OPEN | FLAG_USE | FLAG_REMOVE_LIST
-        | FLAG_CHECK | FLAG_CLEAN | FLAG_SHOW_CONFIG
-        | FLAG_UPDATE_AUTOALIAS))
+        | FLAG_CHECK | FLAG_CLEAN | FLAG_SHOW_CONFIG))
 
     # default config variables
     declare -r CDA_EXEC_NAME_DEFAULT=cda
     declare -r CDA_BASH_COMPLETION_DEFAULT=true
-    declare -r CDA_CMD_FILTER_DEFAULT="percol:peco:fzf:fzy"
+    declare -r CDA_MATCH_EXACT_ORDER_DEFAULT=true
+    declare -r CDA_CMD_FILTER_DEFAULT="peco:percol:fzf:fzy"
     declare -r CDA_CMD_OPEN_DEFAULT="xdg-open:open:ranger:mc"
-    declare -r CDA_CMD_EDITOR_DEFAULT="vim:nano:emacs:vi"
+    declare -r CDA_CMD_EDITOR_DEFAULT="vim:vi:nano:emacs"
     declare -r CDA_ALIAS_MAX_LEN_DEFAULT=16
     declare -r CDA_COLOR_MODE_DEFAULT=auto
     declare -r CDA_BUILTIN_CD_DEFAULT=false
-    declare -r CDA_AUTO_ALIAS_DEFAULT=false
-    declare -r CDA_AUTO_ALIAS_HOOK_TYPE_DEFAULT="$([[ -n ${ZSH_VERSION-} ]] && \printf "chpwd" || \printf "function")"
     declare -r CDA_LIST_HIGHLIGHT_COLOR_DEFAULT="0;0;32"
     declare -r CDA_FILTER_LINE_PREFIX_DEFAULT=true
 
@@ -199,15 +193,6 @@ _cda()
         if [[ $? -ne 0 ]]; then
             return 1
         fi
-    fi
-
-    # -m --multiplexer : temporarily switch to the auto-alias list
-    if _cda::flag::match $FLAG_MULTIPLEXER; then
-        _cda::setup::use --temp "$AUTO_ALIAS_LIST_NAME"
-        if [[ $? -ne 0 ]]; then
-            return 1
-        fi
-        _cda::autoalias::clear
     fi
 
     # -L --list-files
@@ -306,12 +291,6 @@ _cda()
         return $?
     fi
 
-    # --update-autoalias
-    if _cda::flag::match $FLAG_UPDATE_AUTOALIAS; then
-        _cda::autoalias::update
-        return $?
-    fi
-
 
     #------------------------------------------------
     # Actions that do not require aliases
@@ -370,7 +349,15 @@ _cda::setup::init()
         \printf -- "cda: ERROR: Invalid Value: CDA_EXEC_NAME in \"$CDA_DATA_ROOT/config\"\n"
         return 1
     fi
-    eval "$CDA_EXEC_NAME()
+
+    # unalias $CDA_EXEC_NAME to avoid eval error when re-sourcing .bashrc or .zshrc
+    # in case the real alias with the same name of $CDA_EXEC_NAME has been set by alias command.
+    if [[ -n "$(\alias $CDA_EXEC_NAME 2>/dev/null)" ]]; then
+        \unalias "$CDA_EXEC_NAME"
+    fi
+    
+    # define the function with the name of $CDA_EXEC_NAME
+    \eval "$CDA_EXEC_NAME()
     {
         local tmp_argv arg
         tmp_argv=(\"\$@\")
@@ -382,34 +369,6 @@ _cda::setup::init()
         fi
         _cda \"\${tmp_argv[@]}\"
     }"
-
-    # hook cd for the Auto-Alias feature
-    CDA_AUTO_ALIAS="${CDA_AUTO_ALIAS:-$CDA_AUTO_ALIAS_DEFAULT}"
-    CDA_AUTO_ALIAS_HOOK_TYPE="${CDA_AUTO_ALIAS_HOOK_TYPE:-$CDA_AUTO_ALIAS_HOOK_TYPE_DEFAULT}"
-    if [[ -n ${TMUX-} || -n ${STY-} ]]; then
-        if _cda::utils::is_true "$CDA_AUTO_ALIAS"; then
-            case "$CDA_AUTO_ALIAS_HOOK_TYPE" in
-                function)
-                    cd()
-                    {
-                        builtin cd "$@"
-                        [[ $? -ne 0 ]] && return 1
-                        _cda::autoalias::chpwd
-                        return 0
-                    };;
-                prompt)
-                    PROMPT_COMMAND=$(\printf -- "%s" "${PROMPT_COMMAND%;}; _cda::autoalias::chpwd;")
-                    ;;
-                chpwd)
-                    if [[ -n ${ZSH_VERSION-} ]]; then
-                        \autoload -Uz add-zsh-hook
-                        \add-zsh-hook chpwd _cda::autoalias::chpwd
-                    else
-                        \printf -- "cda: WARNING: CDA_AUTO_ALIAS_HOOK_TYPE=chpwd works only in zsh\n"
-                    fi
-            esac
-        fi
-    fi
 
     # configure Bash-Completion
     CDA_BASH_COMPLETION="${CDA_BASH_COMPLETION:-$CDA_BASH_COMPLETION_DEFAULT}"
@@ -426,6 +385,7 @@ _cda::setup::init()
     fi
 
     CDA_FILTER_LINE_PREFIX="${CDA_FILTER_LINE_PREFIX:-$CDA_FILTER_LINE_PREFIX_DEFAULT}"
+    CDA_MATCH_EXACT_ORDER="${CDA_MATCH_EXACT_ORDER:-$CDA_MATCH_EXACT_ORDER_DEFAULT}"
 
     # set initialized flag
     CDA_INITIALIZED=true
@@ -503,19 +463,25 @@ _cda::setup::paths()
         return 1
     fi
 
-    # does the auto-alias list file exist?
-    if [[ ! -e $AUTO_ALIAS_LIST_FILE ]]; then
-        if ! \touch -- "$AUTO_ALIAS_LIST_FILE"; then
-            _cda::msg::error FATAL "Could not create the auto-alias list file: " "$AUTO_ALIAS_LIST_FILE"
+    # does the lastpath file exist?
+    if [[ ! -e $LAST_PATH_FILE ]]; then
+        if ! \touch -- "$LAST_PATH_FILE"; then
+            _cda::msg::error FATAL "Could not create the last path file: " "$LAST_PATH_FILE"
             return 1
         fi
     fi
 
-    # does the auto-pwd file exist?
-    if [[ ! -e $AUTO_ALIAS_PWD_FILE ]]; then
-        if ! \touch -- "$AUTO_ALIAS_PWD_FILE"; then
-            _cda::msg::error FATAL "Could not create the auto-pwd file: " "$AUTO_ALIAS_PWD_FILE"
-            return 1
+    # remove .autolist
+    if [[ -f "$LIST_DIR/.autolist" ]]; then
+        if ! \rm -- "$LIST_DIR/.autolist" 2>/dev/null; then
+            _cda::msg::error WARNING "Could not remove an unnecessary file: " "$LIST_DIR/.autolist"
+        fi
+    fi
+
+    # remove .autopwd
+    if [[ -f "$LIST_DIR/.autopwd" ]]; then
+        if ! \rm -- "$LIST_DIR/.autopwd" 2>/dev/null; then
+            _cda::msg::error WARNING "Could not remove an unnecessary file: " "$LIST_DIR/.autopwd"
         fi
     fi
 
@@ -593,7 +559,7 @@ _cda::setup::use()
     # temporarily change
     if [[ $temp == true ]]; then
         # validate the name of the specified using list 
-        if [[ $list_name != $AUTO_ALIAS_LIST_NAME && ! $list_name =~ ^[a-zA-Z0-9_]+$ ]]; then
+        if [[ ! $list_name =~ ^[a-zA-Z0-9_]+$ ]]; then
             _cda::msg::error ERROR "The list name can not include characters other than \"a-zA-Z0-9_\""
             return 1
         fi
@@ -634,7 +600,7 @@ _cda::setup::use()
         fi
 
         # Write the using list name to the listname file
-        if ! $(printf -- "%s\n" "$list_name" > "$LIST_NAME_FILE"); then
+        if ! $(\printf -- "%s\n" "$list_name" > "$LIST_NAME_FILE"); then
             _cda::msg::error ERROR "Could not write to the file: " "$LIST_DIR_NAME/$list_name"
             return 1
         fi
@@ -654,13 +620,12 @@ _cda::config::create()
 << __EOCFG__ \cat > "$CONFIG_FILE"
 # CDA_EXEC_NAME=$CDA_EXEC_NAME_DEFAULT
 # CDA_BASH_COMPLETION=$CDA_BASH_COMPLETION_DEFAULT
+# CDA_MATCH_EXACT_ORDER=$CDA_MATCH_EXACT_ORDER_DEFAULT
 # CDA_CMD_FILTER=$CDA_CMD_FILTER_DEFAULT
 # CDA_CMD_OPEN=$CDA_CMD_OPEN_DEFAULT
 # CDA_CMD_EDITOR=$CDA_CMD_EDITOR_DEFAULT
 # CDA_FILTER_LINE_PREFIX=$CDA_FILTER_LINE_PREFIX_DEFAULT
 # CDA_BUILTIN_CD=$CDA_BUILTIN_CD_DEFAULT
-# CDA_AUTO_ALIAS=$CDA_AUTO_ALIAS_DEFAULT
-# CDA_AUTO_ALIAS_HOOK_TYPE=$CDA_AUTO_ALIAS_HOOK_TYPE_DEFAULT
 # CDA_COLOR_MODE=$CDA_COLOR_MODE_DEFAULT
 # CDA_LIST_HIGHLIGHT_COLOR="$CDA_LIST_HIGHLIGHT_COLOR_DEFAULT"
 # CDA_ALIAS_MAX_LEN=$CDA_ALIAS_MAX_LEN_DEFAULT
@@ -675,13 +640,12 @@ CDA_SRC_FILE="$CDA_SRC_FILE"
 CDA_DATA_ROOT="$CDA_DATA_ROOT"
 CDA_EXEC_NAME=${CDA_EXEC_NAME:-$CDA_EXEC_NAME_DEFAULT}
 CDA_BASH_COMPLETION=${CDA_BASH_COMPLETION:-$CDA_BASH_COMPLETION_DEFAULT}
+CDA_MATCH_EXACT_ORDER=${CDA_MATCH_EXACT_ORDER:-$CDA_MATCH_EXACT_ORDER_DEFAULT}
 CDA_CMD_FILTER=${CDA_CMD_FILTER:-$CDA_CMD_FILTER_DEFAULT}
 CDA_CMD_OPEN=${CDA_CMD_OPEN:-$CDA_CMD_OPEN_DEFAULT}
 CDA_CMD_EDITOR=${CDA_CMD_EDITOR:-$CDA_CMD_EDITOR_DEFAULT}
 CDA_FILTER_LINE_PREFIX=${CDA_FILTER_LINE_PREFIX:-$CDA_FILTER_LINE_PREFIX_DEFAULT}
 CDA_BUILTIN_CD=${CDA_BUILTIN_CD:-$CDA_BUILTIN_CD_DEFAULT}
-CDA_AUTO_ALIAS=${CDA_AUTO_ALIAS:-$CDA_AUTO_ALIAS_DEFAULT}
-CDA_AUTO_ALIAS_HOOK_TYPE=${CDA_AUTO_ALIAS_HOOK_TYPE:-$CDA_AUTO_ALIAS_HOOK_TYPE_DEFAULT}
 CDA_COLOR_MODE=${CDA_COLOR_MODE:-$CDA_COLOR_MODE_DEFAULT}
 CDA_LIST_HIGHLIGHT_COLOR="${CDA_LIST_HIGHLIGHT_COLOR:-$CDA_LIST_HIGHLIGHT_COLOR_DEFAULT}"
 CDA_ALIAS_MAX_LEN=$CDA_ALIAS_MAX_LEN
@@ -731,7 +695,6 @@ _cda::option::to_flag()
         -p | --path)            \printf -- $FLAG_PATH;;
         -r | --remove)          \printf -- $FLAG_REMOVE;;
         -R | --remove-list)     \printf -- $FLAG_REMOVE_LIST;;
-        -m | --multiplexer)     \printf -- $FLAG_MULTIPLEXER;;
         -s | --subdir)          \printf -- $FLAG_SUBDIR;;
         -u | --use-temp)        \printf -- $FLAG_USE_TEMP;;
         -U | --use)             \printf -- $FLAG_USE;;
@@ -741,7 +704,6 @@ _cda::option::to_flag()
         --list-paths)           \printf -- $FLAG_LIST_PATHS;;
         --config)               \printf -- $FLAG_EDIT_CONFIG;;
         --show-config)          \printf -- $FLAG_SHOW_CONFIG;;
-        --update-autoalias)     \printf -- $FLAG_UPDATE_AUTOALIAS;;
 
         -E | --cmd-editor)      \printf -- $FLAG_NONE;;
         -F | --cmd-filter)      \printf -- $FLAG_NONE;;
@@ -852,14 +814,12 @@ _cda::option::parse()
             -o | --open | \
             -p | --path | \
             -r | --remove | \
-            -m | --multiplexer | \
             -s | --subdir | \
             -v | --verbose | \
             --version | \
             --list-names | \
             --list-paths | \
-            --config | --show-config | \
-            --update-autoalias \
+            --config | --show-config \
             )
                 _cda::option::set $optname || return 1
                 \shift
@@ -909,7 +869,7 @@ _cda::option::parse()
                 done
 
                 # clustered short options requiring NO argument
-                for op in B c C e f h H l L m o p r s v
+                for op in B c C e f h H l L o p r s v
                 do
                     [[ ! $optname =~ ${op} ]] && continue
                     detected_opts="${detected_opts}$op"
@@ -980,16 +940,13 @@ _cda::cd::cd()
     if _cda::flag::match $FLAG_BUILTIN_CD || _cda::utils::is_true "${CDA_BUILTIN_CD-}"; then
         builtin cd "$abs_path"
         rtn=$?
-
-        # instead of the hook function overriding cd command
-        if [[ rtn -eq 0 ]] \
-        && _cda::utils::is_true "${CDA_AUTO_ALIAS}" \
-        && [[ (-n ${TMUX-} || -n ${STY-}) && "$CDA_AUTO_ALIAS_HOOK_TYPE" == "function" ]]; then
-            _cda::autoalias::chpwd
-        fi
     else
         cd "$abs_path"
         rtn=$?
+    fi
+
+    if [[ $rtn -eq 0 ]]; then
+        \pwd > "$LAST_PATH_FILE" 2>/dev/null
     fi
     return $rtn
 }
@@ -1190,8 +1147,15 @@ _cda::list::match()
         [[ -z "$lines" ]] && return 0
 
         if [[ ${#args[@]} -eq 1 ]]; then
-            printf -- "%b\n" "$lines"
+            \printf -- "%b\n" "$lines"
             return 0
+
+        # in an exact order
+        elif _cda::utils::is_true "$CDA_MATCH_EXACT_ORDER"; then
+            regexp="^($(_cda::text::join '[^ ]*' ${args[@]})[^ ]* +)"
+            lines="$(\printf -- "%b" "$lines" | \grep -E "$regexp")"
+
+        # in no particular order
         else
             \unset args[0]
             args=("${args[@]}")
@@ -1199,11 +1163,11 @@ _cda::list::match()
             for name in "${args[@]}"
             do
                 regexp="^${firstArg}[^ ]*${name}[^ ]* +"
-                lines="$(printf -- "%b" "$lines" | \grep -E "$regexp")"
+                lines="$(\printf -- "%b" "$lines" | \grep -E "$regexp")"
                 [[ -z "$lines" ]] && break
-            done 
-            [[ -n "$lines" ]] && printf -- "%b\n" "$lines"
+            done
         fi
+        [[ -n "$lines" ]] && \printf -- "%b\n" "$lines"
     fi
 }
 
@@ -1291,9 +1255,10 @@ _cda::list::path()
         abs_path=$(_cda::dir::select "${@-}" "$Optarg_number")
     
     elif ! _cda::list::is_empty; then
-        # with the last session path
-        if _cda::flag::match $FLAG_MULTIPLEXER && [[ "${1-}" == "-" ]]; then
-            abs_path=$(_cda::autoalias::lastpwd)
+
+        # with the last path
+        if [[ "${1-}" == "-" ]]; then
+            abs_path=$(\cat 2>/dev/null -- "$LAST_PATH_FILE")
 
         # with an alias name
         else
@@ -1671,113 +1636,6 @@ _cda::list::check()
     return 0
 }
 
-
-#------------------------------------------------
-# _cda::autoalias
-#------------------------------------------------
-
-# chpwd is called directly from cd hooked
-_cda::autoalias::chpwd()  
-{
-    [[ -z ${TMUX-} && -z ${STY-} ]] && return 0
-    [[ "$PWD" == "$OLDPWD" ]] && return 0
-    _cda -m --update-autoalias
-    return 0
-}
-
-# --update-autoalias
-_cda::autoalias::update()
-{
-    [[ -z ${TMUX-} && -z ${STY-} ]] && return 1
-
-    local IFS=$' \t\n' line session hostname winpane alias_name
-
-    # tmux
-    if [[ -n ${TMUX-} ]]; then
-        line="$(\tmux display -p "#S:#I_#P")"
-        session="${line%%:*}"
-        winpane="${line#*:}"
-        if ! _cda::alias::validate --name "$session"; then
-            session=$(\sed -e 's/[^a-zA-Z0-9_]/_/g' <<< "$session")
-        fi
-        alias_name="${session}_${winpane}"
-
-    # GNU screen
-    elif [[ -n ${STY-} ]]; then
-        session="${STY#*.}"
-        hostname="$(\hostname)"
-        hostname="${hostname%%.*}"
-        if [[ $session =~ ^[a-z]+-?[0-9]+\.$hostname$ ]]; then
-            session="$(\sed -e 's/^[^0-9]\{1,\}-\{0,1\}\([0-9]\{1,\}\)\..*$/\1/' <<< "$session")"
-            session=$(\printf -- "%d" $((10#${session})))
-        else
-            if ! _cda::alias::validate --name "$session"; then
-                session=$(\sed -e 's/[^a-zA-Z0-9_]/_/g' <<< "$session")
-            fi
-        fi
-        alias_name="${session}_${WINDOW}"
-    fi
-    _cda::flag::set $FLAG_ADD_FORCED
-    _cda::list::add "$alias_name" "$(\pwd)"
-    \pwd > "$AUTO_ALIAS_PWD_FILE"
-}
-
-_cda::autoalias::clear()
-{
-    local IFS=$'\n'
-    local names line session hostname alias_name
-    names=()
-
-    # tmux
-    if _cda::cmd::exist tmux; then
-        while IFS= \read -r line || [[ -n "$line" ]]
-        do
-            session="${line%%:*}"
-            if ! _cda::alias::validate --name "$session"; then
-                session=$(\sed -e 's/[^a-zA-Z0-9_]/_/g' <<< "$session")
-            fi
-            alias_name="${session}_[0-9]+_[0-9]+"
-            names+=("$alias_name")
-        done < <(\tmux ls 2>/dev/null | \grep -E -- '^[^:]+:[ '$'\t'']+[0-9]+[ '$'\t'']+')
-    fi
-
-    # GNU screen
-    if _cda::cmd::exist screen; then
-        hostname="$(\hostname)"
-        hostname="${hostname%%.*}"     
-        while IFS= \read -r line || [[ -n $line ]]
-        do
-            line=$(_cda::text::trim "$line")
-            session="${line#*.}"
-            if [[ $session =~ ^[a-z]+-?[0-9]+\.$hostname$ ]]; then
-                session=$(\sed -e 's/^[^0-9]\{1,\}-\{0,1\}\([0-9]\{1,\}\)\..*$/\1/' <<< "$session")
-                session=$(\printf -- "%d" $((10#${session})))
-            else
-                if ! _cda::alias::validate --name "$session"; then
-                    session=$(\sed -e 's/[^a-zA-Z0-9_]/_/g' <<< "$session")
-                fi
-            fi
-            alias_name="${session}_[0-9]+"
-            names+=("$alias_name")
-        done < <(\screen -ls 2>/dev/null | \grep -E -- '^[ '$'\t'']+[0-9]+\.' | \sed -e 's/(.*)$//')
-    fi
-    if [[ ${#names[*]} -eq 0 ]]; then
-        [[ -s $USING_LIST_FILE ]] && : > "$USING_LIST_FILE"
-        return $?
-    fi
-    
-    local regexp="^($(_cda::text::join ' +|' ${names[@]}) +)"
-    local contents="$(_cda::list::print | \tr "\t" " ")"
-    \grep -E "$regexp" <<< "$contents" | \awk 'NF' | \sort -u > "$USING_LIST_FILE"
-    _cda::utils::check_pipes
-}
-
-_cda::autoalias::lastpwd()
-{
-    \cat 2>/dev/null -- "$AUTO_ALIAS_PWD_FILE"
-}
-
-
 #------------------------------------------------
 # _cda::utils
 #------------------------------------------------
@@ -1969,7 +1827,7 @@ _cda::dir::select()
     local new_abs_path= line=
     
     if _cda::num::is_number "${2-}"; then
-        local dirnum="$(printf -- "%d" $((10#${2})))"
+        local dirnum="$(\printf -- "%d" $((10#${2})))"
         if [[ $dirnum -eq 0 ]]; then
             new_abs_path=$abs_path
         else
@@ -1982,7 +1840,7 @@ _cda::dir::select()
         fi
 
         if _cda::flag::match $FLAG_VERBOSE; then
-            printf -- "%-${CDA_ALIAS_MAX_LEN}s %s\n" "$dirnum" "$new_abs_path" | _cda::list::highlight | _cda::msg::filter 2 >&2
+            \printf -- "%-${CDA_ALIAS_MAX_LEN}s %s\n" "$dirnum" "$new_abs_path" | _cda::list::highlight | _cda::msg::filter 2 >&2
         fi
     else
         # count the number of lines to determine padding width
@@ -2011,7 +1869,7 @@ _cda::dir::select()
 
         if _cda::flag::match $FLAG_VERBOSE; then
             local num="$(<<< "$line" \sed 's/^\([0-9]\{1,\}\).*/\1/')"
-            printf -- "%-${CDA_ALIAS_MAX_LEN}s %s\n" "$num" "$new_abs_path" | _cda::list::highlight | _cda::msg::filter 2 >&2
+            \printf -- "%-${CDA_ALIAS_MAX_LEN}s %s\n" "$num" "$new_abs_path" | _cda::list::highlight | _cda::msg::filter 2 >&2
         fi
     fi
 
@@ -2253,7 +2111,7 @@ _cda::msg::internal_error()
     _cda::msg::error "INTERNAL ERROR" "${funcs[1]}(): $text" "$value"
     funcs[0]= 
     local IFS=$'\n'
-    printf -- "Stack Trace%b\n" "${funcs[*]}"
+    \printf -- "Stack Trace%b\n" "${funcs[*]}"
 }
 
 
@@ -2298,7 +2156,7 @@ _cda::cmd::get()
         *) return 1;;
     esac
     [[ -z $cmd ]] && return 1
-    printf -- "%b" "$cmd"
+    \printf -- "%b" "$cmd"
 }
 
 _cda::cmd::exec()
@@ -2380,7 +2238,6 @@ URL         : https://github.com/itmst71/cda
 Required    : Bash 3.2+ / Zsh 5.0+, Some POSIX commands
 Optional    : Interactive Filter(percol, peco, fzf, fzy etc...)
             : Bash-completion
-            : Terminal Multiplexer(tmux or GNU screen)
 Description : Executes cd with an alias name.
               Supports interactive filters and bash-completion.
 ------------------------------------------------------------------------
@@ -2402,6 +2259,8 @@ ALIAS NAME
     An alias name can include only "a-zA-Z0-9_".
     The default max length is 16.
     e.g.) foo, boo_bar, baz001, a, A, 0, _
+
+    A special alias "-" points the path used last time.
 
 OPTIONS
     --version
@@ -2463,10 +2322,6 @@ OPTIONS
     -C | --clean [ALIAS_NAME [ALIAS_NAME ...]]
         Check the existence of alias paths and remove them
         that do not exist. Also re-format wrong format lines.
-
-    -m | --multiplexer
-        Temporarily switch to auto alias list for terminal
-        multiplexer.
 
     -s | --subdir
         Select a subdirectory in the alias path with a filter.
@@ -2548,9 +2403,16 @@ CONFIG VARIABLES
 
     [ CDA_BASH_COMPLETION ]
         Set to false if you don't want to use Bash-Completion.
-        The defalt is true.
+        The default is true.
 
             CDA_BASH_COMPLETION=true
+
+    [ CDA_MATCH_EXACT_ORDER ]
+        Set to false if you want to use the second and subsequent
+        arguments in no particular order for partial match search.
+        The default is true.
+
+            CDA_MATCH_EXACT_ORDER=$CDA_MATCH_EXACT_ORDER_DEFAULT
 
     [ CDA_CMD_FILTER ]
         Specify the name or path of interactive filter commands
@@ -2586,20 +2448,6 @@ CONFIG VARIABLES
         -B | --builtin-cd override this and temporarily set to true.
 
             CDA_BUILTIN_CD=false
-
-    [ CDA_AUTO_ALIAS ]
-        Set to false if you don't want to use the auto-alias in terminal
-        multiplexer feature. The defalt is true.
-
-            CDA_AUTO_ALIAS=true
-
-    [ CDA_AUTO_ALIAS_HOOK_TYPE ]
-        Specify one of [function prompt chpwd] as the hook type of
-        cd for the auto-alias in terminal multiplexer feature.
-        Only Zsh can use chpwd.
-
-            CDA_AUTO_ALIAS_HOOK_TYPE=function     # Bash
-            CDA_AUTO_ALIAS_HOOK_TYPE=chpwd        # Zsh
 
     [ CDA_COLOR_MODE ]
         Specify one of [never always auto] as the color mode 
@@ -2704,7 +2552,6 @@ _cda::completion::exec()
                 -p --path
                 -r --remove
                 -R --remove-list
-                -m --multiplexer
                 -s --subdir
                 -U --use
                 -u --use-temp
@@ -2737,21 +2584,15 @@ _cda::completion::exec()
 
     # complements an alias name
     else
-        # -m --multiplexer
-        if [[ $COMP_LINE =~ \ (-[a-zA-Z0-9]*m[a-zA-Z0-9]*|--multiplexer)\  ]]; then
-            [[ $COMP_LINE =~ \ [a-zA-Z0-9_]+\  ]] && return 0
-            COMPREPLY=($(\compgen -W "$(_cda -m --list-names  2>/dev/null)" -- "$curr"))
-            return $?
-
         # -u use-temp
-        elif [[ $COMP_LINE =~ \ (-[a-zA-Z0-9]*u|--use-temp)\ +[a-zA-Z0-9_]+\ + ]]; then
-            local using="$(\sed -e 's/^.* \{1,\}\(-[a-zA-Z0-9]*u\|--use-temp\) \{1,\}\([a-zA-Z0-9_]+\) \{1,\}.*$/\2/' <<< "$COMP_LINE")"
+        if [[ $COMP_LINE =~ \ (-[a-zA-Z0-9]*u|--use-temp)\ +[a-zA-Z0-9_]+\ + ]]; then
+            local using="$(\sed -e 's/^.* \{1,\}\(-[a-zA-Z0-9]*u\|--use-temp\) \{1,\}\([a-zA-Z0-9_]\{1,\}\) \{1,\}.*$/\2/' <<< "$COMP_LINE")"
             COMPREPLY=($(\compgen -W "$(_cda -u $using --list-names  2>/dev/null)" -- "$curr"))
             return $?
 
         # aliases in the current list
         else
-            COMPREPLY=($(\compgen -W "$(_cda --list-names  2>/dev/null)" -- "$curr"))
+            COMPREPLY=($(\compgen -W "$(_cda --list-names 2>/dev/null)" -- "$curr"))
             return $?
         fi
     fi